refactor: hide implementation details by making modules "private"
