+++ setup.cfg
@@ -55,7 +55,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    eric = molgenis.bbmri_eric.cli:run
+    eric = molgenis.bbmri_eric._cli:run
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
